feat: show files to be created/replaced
